+++ src/app.js
@@ -11,7 +11,7 @@ fillDatabase();
 
 app.use(logger('dev'));
 app.use((req, res, next) => {
-    res.header("Access-Control-Allow-Origin", 'http://localhost:3000'); // update to match the domain you will make the request from
+    res.header("Access-Control-Allow-Origin", '*'); // update to match the domain you will make the request from
     res.header("Access-Control-Allow-Credentials", "true");
     res.header(
         "Access-Control-Allow-Headers",
