+++ src/routes/images.js
@@ -1,6 +1,6 @@
 const {Router}=require('express') 
 const {postImgCloudinary}=require('../controlers/cloudinary/cloudinary');
-const upload = require('../../multer')
+const upload = require('../lib/multer')
 const router=Router()
 
 router.post('/',upload.array('image',4),postImgCloudinary)
